e2e: Move config types to config package
We move config types to the typs package to avoid import cycles while
refactoring the e2e package. Now we are in better shape and we can keep
the config types and code in the config package.

This means that some code using types.Config need to depend now on the
config package, but this package has no dependencies now, so this is not
an issue.

This change will help to support a new generic config for ramenctl
commands.
