+++ e2e/config/config.go
@@ -5,6 +5,7 @@ package config
 
 import (
 	"fmt"
+	"maps"
 	"regexp"
 	"slices"
 	"strings"
@@ -12,8 +13,6 @@ import (
 	"github.com/spf13/viper"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/validation"
-
-	"github.com/ramendr/ramen/e2e/types"
 )
 
 const (
@@ -35,6 +34,60 @@ const (
 	defaultClusterSetName = "default"
 )
 
+// Channel defines the name and namespace for the channel CR.
+// This is not user-configurable and always uses default values.
+type Channel struct {
+	Name      string `json:"name"`
+	Namespace string `json:"namespace"`
+}
+
+// Namespaces are determined by distro and are not user-configurable.
+type Namespaces struct {
+	RamenHubNamespace       string `json:"ramenHubNamespace"`
+	RamenDRClusterNamespace string `json:"ramenDRClusterNamespace"`
+	RamenOpsNamespace       string `json:"ramenOpsNamespace"`
+	ArgocdNamespace         string `json:"argocdNamespace"`
+}
+
+// Repo represents the user-configurable git repository settings.
+// It includes the repository url and branch to be used for deploying workload.
+type Repo struct {
+	URL    string `json:"url"`
+	Branch string `json:"branch"`
+}
+
+type PVCSpec struct {
+	Name             string `json:"name"`
+	StorageClassName string `json:"storageClassName"`
+	AccessModes      string `json:"accessModes"`
+}
+
+type Cluster struct {
+	Kubeconfig string `json:"kubeconfig"`
+}
+
+type Test struct {
+	Workload string `json:"workload"`
+	Deployer string `json:"deployer"`
+	PVCSpec  string `json:"pvcSpec"`
+}
+
+// Config keeps configuration for e2e tests.
+type Config struct {
+	// User configurable values.
+	Clusters   map[string]Cluster `json:"clusters"`
+	ClusterSet string             `json:"clusterSet"`
+	Distro     string             `json:"distro"`
+	Repo       Repo               `json:"repo"`
+	DRPolicy   string             `json:"drPolicy"`
+	PVCSpecs   []PVCSpec          `json:"pvcSpecs"`
+	Tests      []Test             `json:"tests"`
+
+	// Generated values
+	Channel    Channel    `json:"channel"`
+	Namespaces Namespaces `json:"namespaces"`
+}
+
 // Options that can be used in a configuration file.
 type Options struct {
 	Workloads []string
@@ -42,7 +95,7 @@ type Options struct {
 }
 
 // Default namespace mappings for Kubernetes (k8s) clusters.
-var K8sNamespaces = types.NamespacesConfig{
+var K8sNamespaces = Namespaces{
 	RamenHubNamespace:       "ramen-system",
 	RamenDRClusterNamespace: "ramen-system",
 	RamenOpsNamespace:       "ramen-ops",
@@ -50,7 +103,7 @@ var K8sNamespaces = types.NamespacesConfig{
 }
 
 // Default namespace mappings for OpenShift (ocp) clusters.
-var OcpNamespaces = types.NamespacesConfig{
+var OcpNamespaces = Namespaces{
 	RamenHubNamespace:       "openshift-operators",
 	RamenDRClusterNamespace: "openshift-dr-system",
 	RamenOpsNamespace:       "openshift-dr-ops",
@@ -59,8 +112,8 @@ var OcpNamespaces = types.NamespacesConfig{
 
 var resourceNameForbiddenCharacters *regexp.Regexp
 
-func ReadConfig(configFile string, options Options) (*types.Config, error) {
-	config := &types.Config{}
+func ReadConfig(configFile string, options Options) (*Config, error) {
+	config := &Config{}
 
 	if err := readConfig(configFile, config); err != nil {
 		return nil, err
@@ -88,7 +141,7 @@ func ReadConfig(configFile string, options Options) (*types.Config, error) {
 	return config, nil
 }
 
-func readConfig(configFile string, config *types.Config) error {
+func readConfig(configFile string, config *Config) error {
 	viper.SetDefault("Repo.URL", defaultGitURL)
 	viper.SetDefault("Repo.Branch", defaultGitBranch)
 	viper.SetDefault("DRPolicy", defaultDRPolicyName)
@@ -107,7 +160,7 @@ func readConfig(configFile string, config *types.Config) error {
 	return nil
 }
 
-func validateDistro(config *types.Config) error {
+func validateDistro(config *Config) error {
 	// Discover distro during validation if the distro is not configured
 	if config.Distro == "" {
 		return nil
@@ -126,7 +179,7 @@ func validateDistro(config *types.Config) error {
 	return nil
 }
 
-func validateClusters(config *types.Config) error {
+func validateClusters(config *Config) error {
 	if config.Clusters["hub"].Kubeconfig == "" {
 		return fmt.Errorf("failed to find hub cluster in configuration")
 	}
@@ -142,7 +195,7 @@ func validateClusters(config *types.Config) error {
 	return nil
 }
 
-func validatePVCSpecs(config *types.Config) error {
+func validatePVCSpecs(config *Config) error {
 	if len(config.PVCSpecs) == 0 {
 		return fmt.Errorf("failed to find pvcs in configuration")
 	}
@@ -167,8 +220,8 @@ func validatePVCSpecs(config *types.Config) error {
 }
 
 // validateDuplicatePVCSpecsNames ensures no PVCSpec has a duplicate name
-func validateDuplicatePVCSpecsNames(pvcSpecs []types.PVCSpecConfig) error {
-	seen := make(map[string]types.PVCSpecConfig)
+func validateDuplicatePVCSpecsNames(pvcSpecs []PVCSpec) error {
+	seen := make(map[string]PVCSpec)
 	for _, spec := range pvcSpecs {
 		if existing, exists := seen[spec.Name]; exists {
 			return fmt.Errorf("duplicate pvcSpec name %q found:\n	%+v\n	%+v", spec.Name, existing, spec)
@@ -181,11 +234,11 @@ func validateDuplicatePVCSpecsNames(pvcSpecs []types.PVCSpecConfig) error {
 }
 
 // validateDuplicatePVCSpecsContent ensures no two PVCSpecs have the same storageClassName and accessModes
-func validateDuplicatePVCSpecsContent(pvcSpecs []types.PVCSpecConfig) error {
-	seen := make(map[types.PVCSpecConfig]types.PVCSpecConfig)
+func validateDuplicatePVCSpecsContent(pvcSpecs []PVCSpec) error {
+	seen := make(map[PVCSpec]PVCSpec)
 
 	for _, spec := range pvcSpecs {
-		key := types.PVCSpecConfig{
+		key := PVCSpec{
 			StorageClassName: spec.StorageClassName,
 			AccessModes:      spec.AccessModes,
 		}
@@ -204,7 +257,7 @@ func validateDuplicatePVCSpecsContent(pvcSpecs []types.PVCSpecConfig) error {
 // conforms to Kubernetes DNS subdomain naming rules (RFC 1123).
 // https://kubernetes.io/docs/concepts/overview/working-with-objects/names/#dns-subdomain-names
 // Returns an error if any StorageClassName in pvcSpec is invalid.
-func validateStorageClassNameFormat(pvcSpecs []types.PVCSpecConfig) error {
+func validateStorageClassNameFormat(pvcSpecs []PVCSpec) error {
 	for _, spec := range pvcSpecs {
 		errs := validation.NameIsDNSSubdomain(spec.StorageClassName, false)
 		if len(errs) > 0 {
@@ -217,7 +270,7 @@ func validateStorageClassNameFormat(pvcSpecs []types.PVCSpecConfig) error {
 }
 
 // validateAccessModes validates that accessModes is one of the supported values
-func validateAccessModes(pvcSpecs []types.PVCSpecConfig) error {
+func validateAccessModes(pvcSpecs []PVCSpec) error {
 	validModes := map[corev1.PersistentVolumeAccessMode]struct{}{
 		corev1.ReadWriteOnce:    {},
 		corev1.ReadOnlyMany:     {},
@@ -235,7 +288,7 @@ func validateAccessModes(pvcSpecs []types.PVCSpecConfig) error {
 	return nil
 }
 
-func validateTests(config *types.Config, options *Options) error {
+func validateTests(config *Config, options *Options) error {
 	// We allow an empty test list so one can run the validation tests or unit tests without a fully configured file.
 	if len(config.Tests) == 0 {
 		return nil
@@ -246,7 +299,7 @@ func validateTests(config *types.Config, options *Options) error {
 		pvcSpecNames = append(pvcSpecNames, spec.Name)
 	}
 
-	testsSeen := map[types.TestConfig]struct{}{}
+	testsSeen := map[Test]struct{}{}
 
 	for _, t := range config.Tests {
 		if _, ok := testsSeen[t]; ok {
@@ -273,13 +326,60 @@ func validateTests(config *types.Config, options *Options) error {
 }
 
 // PVCSpecMap returns a mapping from PVCSpec.Name to PVCSpec.
-func PVCSpecsMap(config *types.Config) map[string]types.PVCSpecConfig {
-	res := map[string]types.PVCSpecConfig{}
+func PVCSpecsMap(config *Config) map[string]PVCSpec {
+	res := map[string]PVCSpec{}
 	for _, spec := range config.PVCSpecs {
 		res[spec.Name] = spec
 	}
 
 	return res
+}
+
+// Equal return true if config is equal to other config.
+//
+//nolint:cyclop
+func (c *Config) Equal(o *Config) bool {
+	if c == o {
+		return true
+	}
+
+	if c.Distro != o.Distro {
+		return false
+	}
+
+	if c.Repo != o.Repo {
+		return false
+	}
+
+	if c.DRPolicy != o.DRPolicy {
+		return false
+	}
+
+	if c.ClusterSet != o.ClusterSet {
+		return false
+	}
+
+	if !maps.Equal(c.Clusters, o.Clusters) {
+		return false
+	}
+
+	if !slices.Equal(c.PVCSpecs, o.PVCSpecs) {
+		return false
+	}
+
+	if !slices.Equal(c.Tests, o.Tests) {
+		return false
+	}
+
+	if c.Channel != o.Channel {
+		return false
+	}
+
+	if c.Namespaces != o.Namespaces {
+		return false
+	}
+
+	return true
 }
 
 // resourceName convert a URL to conventional k8s resource name:
